% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.empty.R
\name{as.empty}
\alias{as.empty}
\alias{as.empty.default}
\alias{as.empty.data.frame}
\title{Convert to an empty object}
\usage{
as.empty(x, ...)

\method{as.empty}{default}(x, ...)

\method{as.empty}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional parameterS.}
}
\value{
An empty object
}
\description{
Convert \code{x} to an empty object.
}
\examples{
x <- c("a", "b", "c")
as.empty(x)
class(as.empty(x)) # still a character

x <- factor(LETTERS)
as.empty(x)        # levels are kept
class(as.empty(x)) # still a factor

x <- data.frame(x = 1:3, y = 2:4)
as.empty(x)

}
\seealso{
\code{\link{is.empty}} in this package.
}
