% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_empty.R
\name{as_empty}
\alias{as_empty}
\alias{as.empty}
\alias{as_empty.default}
\alias{as_empty.data.frame}
\title{Convert to an empty object}
\usage{
as_empty(x, ...)

as.empty(x, ...)

\method{as_empty}{default}(x, ...)

\method{as_empty}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional parameterS.}
}
\value{
An empty object
}
\description{
Convert \code{x} to an empty object.
}
\examples{
x <- c("a", "b", "c")
as_empty(x)
class(as_empty(x)) # still a character

x <- factor(LETTERS)
as_empty(x)        # levels are kept
class(as_empty(x)) # still a factor

x <- data.frame(x = 1:3, y = 2:4)
as_empty(x)

}
\seealso{
\code{\link[bazar]{is_empty}} in this package.
}
