% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copy.R
\name{Copy}
\alias{Copy}
\title{Copy data to the clipboard}
\usage{
Copy(x, size = 128L, quote = TRUE, sep = "\\t", na = "",
  dec = ".", ...)
}
\arguments{
\item{x}{An object.}

\item{size}{integer. Number of kilobytes. 
Increase this value if the object \code{x} 
is too big.}

\item{quote}{See the eponymous argument in \code{\link[utils]{write.table}}.}

\item{sep}{character. The field separator string.}

\item{na}{character. The string to use for missing values.}

\item{dec}{character. The string to use for decimal points 
in numeric or complex columns.}

\item{...}{Additional arguments to be passed to 
\code{\link[utils]{write.table}}.}
}
\description{
The function \code{Copy} can typically be used 
to copy data from a data frame, in order to paste 
it somewhere else (in Excel for instance).
}
