% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erase.R
\name{erase}
\alias{erase}
\title{Delete objects}
\usage{
erase(ask = TRUE)
}
\arguments{
\item{ask}{logical. If \code{TRUE} (the default), a confirmation is interactively 
asked to the user.}
}
\description{
The function \code{erase} deletes all objects 
that live in the calling environment.
}
\section{Warning}{
 use this function with care!
}

