% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_funs.R
\name{get_all_funs}
\alias{get_all_funs}
\title{Functions exported by a package}
\usage{
get_all_funs(pkg)
}
\arguments{
\item{pkg}{character. The package of interest. (Must be installed already.)}
}
\value{
A character vector, the functions exported.
}
\description{
\code{get_all_funs} provides all the functions exported by a given 
installed package.
}
\examples{
get_all_funs("stats")

}
