% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_pkgs.R
\name{get_all_pkgs}
\alias{get_all_pkgs}
\title{Packages exporting a function}
\usage{
get_all_pkgs(fun, packages = NULL)
}
\arguments{
\item{fun}{function or character. The function of interest.}

\item{packages}{The packages to look into. If \code{NULL}, the list of currently attached
packages is explored.}
}
\value{
A character vector, the packages.
}
\description{
\code{get_all_pkgs} provides all packages (belonging to a given list of 
packages) exported by a given function.
}
\examples{
\dontrun{
get_all_pkgs("as.fun")
get_all_pkgs(as.fun)
get_all_pkgs("stats::median")
}

}
