% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Value matching}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{table}{vector or NULL: the values to be matched against.}
}
\value{
A logical vector, indicating if a non-match was located for each element of x:
thus the values are TRUE or FALSE and never NA.
}
\description{
The function \code{\%nin\%} is the negation 
of the function \code{\%in\%}.
}
\examples{
1:10 \%nin\% c(1,3,5,9)

}
\seealso{
\code{\link{match}}.
}
