% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.empty.R
\name{is.empty}
\alias{is.empty}
\alias{is.empty.default}
\alias{is.empty.data.frame}
\title{Test emptyness}
\usage{
is.empty(x)

\method{is.empty}{default}(x)

\method{is.empty}{data.frame}(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
\code{TRUE} if \code{x} is empty, \code{FALSE} otherwise.
}
\description{
These methods test if an object \code{x} is empty.
}
\examples{
is.empty(4)
is.empty(c())
is.empty(new.env())
is.empty(character(0))
is.empty(list())
is.empty(integer(0))
is.empty(data.frame())

}
\seealso{
\code{\link[bazar]{as.empty}} in this package.
}
