% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.formula.R
\name{is.formula}
\alias{is.formula}
\title{Test if an object is a formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical, \code{TRUE} if \code{x} is a formula.
}
\description{
The function \code{is.formula} tests if the object 
\code{x} is a formula.
}
\examples{
is.formula("this is a formula")
is.formula(f <- formula("y ~ x"))
is.formula(update(f, ~ . -1))

}
