% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\alias{is.empty}
\alias{is_empty.default}
\alias{is_empty.data.frame}
\title{Test emptyness}
\usage{
is_empty(x)

is.empty(x)

\method{is_empty}{default}(x)

\method{is_empty}{data.frame}(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
\code{TRUE} if \code{x} is empty, \code{FALSE} otherwise.
}
\description{
These methods test if an object \code{x} is empty.
}
\examples{
is_empty(4)
is_empty(c())
is_empty(new.env())
is_empty(character(0))
is_empty(list())
is_empty(integer(0))
is_empty(data.frame())

}
\seealso{
\code{\link[bazar]{as_empty}} in this package.
}
