% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_formula.R
\name{is_formula}
\alias{is_formula}
\alias{is.formula}
\title{Test if an object is a formula}
\usage{
is_formula(x)

is.formula(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical, \code{TRUE} if \code{x} is a formula.
}
\description{
The function \code{is_formula} tests if the object 
\code{x} is a formula.
}
\examples{
is_formula("this is a formula")
is_formula(f <- formula("y ~ x"))
is_formula(update(f, ~ . -1))

}
