% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNA.R
\name{isNA}
\alias{isNA}
\title{Test if NA}
\usage{
isNA(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{isNA} tests if an object \code{x} is identical to one of \code{NA}, 
\code{NA_character_}, \code{NA_complex_}, \code{NA_integer_}, 
\code{NA_real_}, or \code{NaN}.
}
\seealso{
\code{\link{isTRUE}}.
}
