% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlist.R
\name{nlist}
\alias{nlist}
\alias{as.nlist}
\alias{is.nlist}
\title{Named lists}
\usage{
nlist(...)

as.nlist(x, ...)

is.nlist(x)
}
\arguments{
\item{...}{Named objects.}

\item{x}{Object to be coerced or tested.}
}
\value{
A named list.
}
\description{
Functions to construct, coerce and check for named lists.
}
\examples{
x <- nlist(x = 2, y = c("a", "b"))
is.nlist(x)

}
