% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause.R
\name{pause}
\alias{pause}
\title{Have a rest, make a pause}
\usage{
pause(duration = Inf)
}
\arguments{
\item{duration}{numeric or infinite. 
If \code{duration} is infinite (the default), then a pause is 
made until the user presses <Enter> or types 'stop'. 
Else if \code{x = duration} is a number, then a pause is made 
during \code{x} seconds.}
}
\description{
The \code{pause} function stops momentarily the 
execution of a program. 
Pressing <Enter> continues the execution; typing 'stop' 
(without quotation marks) ends the program.
}
\seealso{
\code{\link{Sys.sleep}}.
}
