% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle2.R
\name{rle2}
\alias{rle2}
\title{Run length encoding (modified version)}
\usage{
rle2(x, tolerance = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{numeric vector.}

\item{tolerance}{numeric. Differences smaller than tolerance are considered as equal.
The default value is close to \code{1.5e-8}.}
}
\value{
An object of class \code{"rle"} which is a list with components:
\itemize{
\item \code{lengths}: an integer vector containing the length of each run.
\item \code{values}: a vector of the same length as lengths with the 
corresponding values.
}
}
\description{
Compute the lengths and values of runs of 
\code{\link[bazar]{almost.equal}} values in a vector.
}
\seealso{
\code{\link[bazar]{almost.equal}} in this package; 
\code{\link[base]{rle}} in package \pkg{base}.
}
