% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionPackages.R
\name{sessionPackages}
\alias{sessionPackages}
\title{Shows packages attached to the current R session}
\usage{
sessionPackages(package = NULL)
}
\arguments{
\item{package}{a character vector naming installed packages,
or \code{NULL} (the default) meaning all attached packages.}
}
\value{
A list with the following components:
\itemize{
  \item \code{basePkgs}: a character vector of base packages which are attached.
  \item \code{otherPkgs} (not always present): a character vector of other attached packages.
}
}
\description{
The function \code{sessionPackages} prints the list
of packages attached to the current R session.
}
\details{
This function reuses part of the code from 
\code{\link[utils]{sessionInfo}}.
}
\examples{
sessionPackages()

}
\seealso{
\code{\link{sessionInfo}} from package \pkg{utils}, 
\code{\link{R.version}} from package \pkg{base}.
}
