% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_bot.R
\name{top}
\alias{top}
\alias{bot}
\title{Top or bottom element of an object}
\usage{
top(x)

bot(x)
}
\arguments{
\item{x}{an object.}
}
\value{
An object (usually) like \code{x} but generally smaller.
}
\description{
\code{top(x)} is an alias for \code{head(x, 1L)}. 
\code{bot(x)} is an alias for \code{tail(x, 1L)}.
}
\seealso{
\code{\link[utils]{head}} and \code{\link[utils]{head}} from package 
\pkg{utils}
}
