% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Removes extra whitespaces from a string}
\usage{
trim(x)
}
\arguments{
\item{x}{character. The character vector at stake.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
The function \code{trim} removes unnecessary whitespaces
from a character vector.
}
\examples{
trim(c(" a b", "Hello  World "))

}
\seealso{
\code{\link{gsub}}.
}
