% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verlan.R
\name{verlan}
\alias{verlan}
\title{Back slang}
\usage{
verlan(x)
}
\arguments{
\item{x}{character. A vector of strings.}
}
\value{
A character vector of the same length as \code{x}.
}
\description{
The \code{verlan} function reverses the order of 
the characters in a string.
}
\examples{
verlan("baba") ## "abab"
verlan(c("radar", "paul")) ## c("radar", "luap")

}
