% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bazar-package.R
\docType{package}
\name{bazar}
\alias{bazar}
\alias{bazar-package}
\title{bazar: miscellaneous basic functions}
\description{
bazar provides a collection of miscellaneous functions for 
\itemize{
  \item copying objects to the clipboard (\code{\link{Copy}});
  \item manipulating strings (\code{\link{concat}}, \code{\link{mgsub}}, 
  \code{\link{trim}}, \code{\link{verlan}}); 
  \item loading or showing packages (\code{\link{library_with_dep}}, 
  \code{\link{require_with_dep}}, \code{\link{sessionPackages}}); 
  \item creating or testing for named lists (\code{\link{nlist}}, 
  \code{\link{as_nlist}}, \code{\link{is_nlist}}), 
  formulas (\code{\link{is_formula}}), empty objects 
  (\code{\link{as_empty}}, \code{\link{is_empty}}), 
  whole numbers (\code{\link{as_wholenumber}}, 
  \code{\link{is_wholenumber}}); 
  \item testing for equality (\code{\link{almost_equal}}, 
  \code{\link{almost_zero}}); 
  \item getting modified versions of usual functions 
  (\code{\link{rle2}}, \code{\link{sumNA}}); 
  \item making a pause or a stop (\code{\link{pause}}, 
  \code{\link{stopif}}); 
  \item and others (\code{\link{erase}}, \code{\link{\%in\%}}, 
  \code{\link{unwhich}}). 
}
}
